requirements.txt
# Enhanced Emotion Recognition System - Requirements

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Science and Machine Learning
numpy>=1.24.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress and Utilities
tqdm>=4.65.0
pathlib2>=2.3.7

# Data Analysis
pandas>=2.0.0

# Configuration and Logging
pyyaml>=6.0
logging
datetime
json
argparse

# Dimensionality Reduction
scikit-learn>=1.3.0  # For t-SNE and PCA

# Image Processing Enhancement (Optional but recommended)
imageio>=2.31.0
scikit-image>=0.21.0

# Development and Testing (Optional)
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0

# DDColor Dependencies (for image colorization preprocessing)
# Note: DDColor repository should be cloned separately
# Additional dependencies for DDColor:
# timm>=0.9.0
# basicsr>=1.4.2
# facexlib>=0.3.0
# gfpgan>=1.3.8

# SwinIR Dependencies (for image super-resolution preprocessing)
# Note: SwinIR repository should be cloned separately
# Additional dependencies for SwinIR:
# timm>=0.9.0
# basicsr>=1.4.2

# Optional GPU acceleration
# nvidia-ml-py3>=7.352.0  # For GPU monitoring

# Performance monitoring
psutil>=5.9.0

# Web interface (if needed for deployment)
# flask>=2.3.0
# flask-cors>=4.0.0
# gradio>=3.40.0

# Data handling
h5py>=3.9.0

# Experiment tracking (optional)
# wandb>=0.15.0
# tensorboard>=2.13.0

# Production deployment (optional)
# gunicorn>=21.0.0
# uvicorn>=0.23.0

# Documentation (optional)
# sphinx>=7.1.0
# sphinx-rtd-theme>=1.3.0
